Change default compiler to clang
Currently, my g++ produces crashing executable ("general
protection fault", whatever that means).

Therefore, switch to clang for now. Sadly this requires
fixing of a use structured binding as lambda capture case.
Technically, clang is in the right here, the C++ standard
really doesn't allow for that. But from a user's point of
view it could just accept it as g++ does.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>