Reset correct undo buffer on load
When loading into a new window, the undo buffer of the old window
was reset.

The root cause was that the second part of the load function was
not accessing the document object itself, but a pointer to either
itself (load into same window) or to a new document (load into
new window). This is obviously very error-prone.

Therefore, split out the second part of the load function into
a new function.

Signed-off-by: Berthold Stoeger <bstoeger@mail.tuwien.ac.at>